refactor(config): split loader.py into sub-modules
Extract 6 sub-modules from the 3365-line monolith:
- coercers.py: type coercion/normalization helpers (~150 lines)
- resolvers.py: config value resolution (~106 lines)
- admin.py: local admin management (~44 lines)
- webui_settings.py: WebUI settings class (~61 lines)
- model_parsers.py: all model config parsers (~1250 lines)
- domain_parsers.py: domain config parsers + _update_dataclass (~284 lines)

loader.py retains Config class, TOML loading, and all re-exports for
backward compatibility. All 1429 tests pass, mypy strict clean.
