+++ src/Undefined/config/loader.py
@@ -2,7 +2,6 @@
 
 from __future__ import annotations
 
-import json
 import logging
 import os
 import re
@@ -37,38 +36,90 @@ from .models import (
     ImageGenConfig,
     ImageGenModelConfig,
     MemeConfig,
-    ModelPool,
-    ModelPoolEntry,
     NagaConfig,
     RerankModelConfig,
     SecurityModelConfig,
     VisionModelConfig,
 )
-from Undefined.utils.request_params import (
-    merge_request_params,
-    normalize_request_params,
+from .coercers import (  # noqa: F401 — re-exported for backward compat
+    _coerce_bool,
+    _coerce_float,
+    _coerce_int,
+    _coerce_int_list,
+    _coerce_str,
+    _coerce_str_list,
+    _get_model_request_params,
+    _get_value,
+    _normalize_base_url,
+    _normalize_queue_interval,
+    _normalize_str,
+    _warn_env_fallback,
+)
+from .resolvers import (  # noqa: F401 — re-exported for backward compat
+    _resolve_api_mode,
+    _resolve_reasoning_effort,
+    _resolve_reasoning_effort_style,
+    _resolve_responses_force_stateless_replay,
+    _resolve_responses_tool_choice_compat,
+    _resolve_thinking_compat_flags,
+)
+from .admin import (  # noqa: F401 — re-exported for backward compat
+    LOCAL_CONFIG_PATH,
+    load_local_admins,
+    save_local_admins,
+)
+from .webui_settings import (  # noqa: F401 — re-exported for backward compat
+    DEFAULT_WEBUI_PASSWORD,
+    DEFAULT_WEBUI_PORT,
+    DEFAULT_WEBUI_URL,
+    WebUISettings,
+    load_webui_settings,
+)
+from .model_parsers import (
+    _log_debug_info,
+    _merge_admins,
+    _parse_agent_model_config,
+    _parse_chat_model_config,
+    _parse_embedding_model_config,
+    _parse_grok_model_config,
+    _parse_historian_model_config,
+    _parse_image_edit_model_config,
+    _parse_image_gen_config,
+    _parse_image_gen_model_config,
+    _parse_naga_model_config,
+    _parse_rerank_model_config,
+    _parse_security_model_config,
+    _parse_summary_model_config,
+    _parse_vision_model_config,
+    _verify_required_fields,
+)
+from .domain_parsers import (
+    _parse_api_config,
+    _parse_cognitive_config,
+    _parse_easter_egg_call_mode,
+    _parse_memes_config,
+    _parse_naga_config,
+    _update_dataclass,
 )
 
 logger = logging.getLogger(__name__)
 
 CONFIG_PATH = Path("config.toml")
-LOCAL_CONFIG_PATH = Path("config.local.json")
 
-DEFAULT_WEBUI_URL = "127.0.0.1"
-DEFAULT_WEBUI_PORT = 8787
-DEFAULT_WEBUI_PASSWORD = "changeme"
-DEFAULT_API_HOST = "127.0.0.1"
-DEFAULT_API_PORT = 8788
-DEFAULT_API_AUTH_KEY = "changeme"
-
-_ENV_WARNED_KEYS: set[str] = set()
-
-
-def _warn_env_fallback(name: str) -> None:
-    if name in _ENV_WARNED_KEYS:
-        return
-    _ENV_WARNED_KEYS.add(name)
-    logger.warning("检测到环境变量 %s，建议迁移到 config.toml", name)
+# Re-export symbols that external modules import from this module.
+__all__ = [
+    "CONFIG_PATH",
+    "Config",
+    "DEFAULT_WEBUI_PASSWORD",
+    "DEFAULT_WEBUI_PORT",
+    "DEFAULT_WEBUI_URL",
+    "LOCAL_CONFIG_PATH",
+    "WebUISettings",
+    "load_local_admins",
+    "load_toml_data",
+    "load_webui_settings",
+    "save_local_admins",
+]
 
 
 def _load_env() -> None:
@@ -141,308 +192,6 @@ def load_toml_data(
         if strict:
             raise ValueError(str(exc)) from exc
         return {}
-
-
-def _get_nested(data: dict[str, Any], path: tuple[str, ...]) -> Any:
-    node: Any = data
-    for key in path:
-        if not isinstance(node, dict) or key not in node:
-            return None
-        node = node[key]
-    return node
-
-
-def _normalize_str(value: Any) -> Optional[str]:
-    if value is None:
-        return None
-    if isinstance(value, str):
-        stripped = value.strip()
-        return stripped if stripped else None
-    return str(value).strip()
-
-
-def _coerce_int(value: Any, default: int) -> int:
-    if value is None:
-        return default
-    try:
-        return int(value)
-    except (TypeError, ValueError):
-        return default
-
-
-def _coerce_float(value: Any, default: float) -> float:
-    if value is None:
-        return default
-    try:
-        return float(value)
-    except (TypeError, ValueError):
-        return default
-
-
-def _normalize_queue_interval(value: float, default: float = 1.0) -> float:
-    """规范化队列发车间隔。
-
-    `0` 表示立即发车，负数回退到默认值。
-    """
-
-    return default if value < 0 else value
-
-
-def _coerce_bool(value: Any, default: bool) -> bool:
-    if isinstance(value, bool):
-        return value
-    if isinstance(value, (int, float)):
-        return bool(value)
-    if isinstance(value, str):
-        return value.strip().lower() in {"1", "true", "yes", "on"}
-    return default
-
-
-def _coerce_str(value: Any, default: str) -> str:
-    normalized = _normalize_str(value)
-    return normalized if normalized is not None else default
-
-
-def _normalize_base_url(value: str, default: str) -> str:
-    normalized = value.strip().rstrip("/")
-    if normalized:
-        return normalized
-    return default.rstrip("/")
-
-
-def _coerce_int_list(value: Any) -> list[int]:
-    if value is None:
-        return []
-    if isinstance(value, list):
-        items: list[int] = []
-        for item in value:
-            try:
-                items.append(int(item))
-            except (TypeError, ValueError):
-                continue
-        return items
-    if isinstance(value, str):
-        parts = [part.strip() for part in value.split(",") if part.strip()]
-        items = []
-        for part in parts:
-            try:
-                items.append(int(part))
-            except ValueError:
-                continue
-        return items
-    return []
-
-
-def _coerce_str_list(value: Any) -> list[str]:
-    if value is None:
-        return []
-    if isinstance(value, list):
-        return [str(item).strip() for item in value if str(item).strip()]
-    if isinstance(value, str):
-        return [part.strip() for part in value.split(",") if part.strip()]
-    return []
-
-
-def _coerce_request_params(value: Any) -> dict[str, Any]:
-    return normalize_request_params(value)
-
-
-def _get_model_request_params(data: dict[str, Any], model_name: str) -> dict[str, Any]:
-    return _coerce_request_params(
-        _get_nested(data, ("models", model_name, "request_params"))
-    )
-
-
-def _get_value(
-    data: dict[str, Any],
-    path: tuple[str, ...],
-    env_key: Optional[str],
-) -> Any:
-    value = _get_nested(data, path)
-    if value is not None:
-        return value
-    if env_key:
-        env_value = os.getenv(env_key)
-        if env_value is not None and str(env_value).strip() != "":
-            _warn_env_fallback(env_key)
-            return env_value
-    return None
-
-
-_VALID_API_MODES = {"chat_completions", "responses"}
-_VALID_REASONING_EFFORT_STYLES = {"openai", "anthropic"}
-
-
-def _resolve_reasoning_effort_style(value: Any, default: str = "openai") -> str:
-    style = _coerce_str(value, default).strip().lower()
-    if style not in _VALID_REASONING_EFFORT_STYLES:
-        return default
-    return style
-
-
-def _resolve_thinking_compat_flags(
-    data: dict[str, Any],
-    model_name: str,
-    include_budget_env_key: str,
-    tool_call_compat_env_key: str,
-    legacy_env_key: str,
-) -> tuple[bool, bool]:
-    """解析思维链兼容配置，并兼容旧字段 deepseek_new_cot_support。"""
-    include_budget_value = _get_value(
-        data,
-        ("models", model_name, "thinking_include_budget"),
-        include_budget_env_key,
-    )
-    tool_call_compat_value = _get_value(
-        data,
-        ("models", model_name, "thinking_tool_call_compat"),
-        tool_call_compat_env_key,
-    )
-    legacy_value = _get_value(
-        data,
-        ("models", model_name, "deepseek_new_cot_support"),
-        legacy_env_key,
-    )
-
-    include_budget_default = True
-    tool_call_compat_default = True
-    if legacy_value is not None:
-        legacy_enabled = _coerce_bool(legacy_value, False)
-        include_budget_default = not legacy_enabled
-        tool_call_compat_default = legacy_enabled
-
-    return (
-        _coerce_bool(include_budget_value, include_budget_default),
-        _coerce_bool(tool_call_compat_value, tool_call_compat_default),
-    )
-
-
-def _resolve_api_mode(
-    data: dict[str, Any],
-    model_name: str,
-    env_key: str,
-    default: str = "chat_completions",
-) -> str:
-    raw_value = _get_value(data, ("models", model_name, "api_mode"), env_key)
-    value = _coerce_str(raw_value, default).strip().lower()
-    if value not in _VALID_API_MODES:
-        return default
-    return value
-
-
-def _resolve_reasoning_effort(value: Any, default: str = "medium") -> str:
-    return _coerce_str(value, default).strip().lower()
-
-
-def _resolve_responses_tool_choice_compat(
-    data: dict[str, Any],
-    model_name: str,
-    env_key: str,
-    default: bool = False,
-) -> bool:
-    return _coerce_bool(
-        _get_value(
-            data,
-            ("models", model_name, "responses_tool_choice_compat"),
-            env_key,
-        ),
-        default,
-    )
-
-
-def _resolve_responses_force_stateless_replay(
-    data: dict[str, Any],
-    model_name: str,
-    env_key: str,
-    default: bool = False,
-) -> bool:
-    return _coerce_bool(
-        _get_value(
-            data,
-            ("models", model_name, "responses_force_stateless_replay"),
-            env_key,
-        ),
-        default,
-    )
-
-
-def load_local_admins() -> list[int]:
-    """从本地配置文件加载动态管理员列表"""
-    if not LOCAL_CONFIG_PATH.exists():
-        return []
-    try:
-        with open(LOCAL_CONFIG_PATH, "r", encoding="utf-8") as f:
-            data = json.load(f)
-        admin_qqs: list[int] = data.get("admin_qqs", [])
-        return admin_qqs
-    except Exception as exc:
-        logger.warning("读取本地配置失败: %s", exc)
-        return []
-
-
-def save_local_admins(admin_qqs: list[int]) -> None:
-    """保存动态管理员列表到本地配置文件"""
-    try:
-        data: dict[str, list[int]] = {}
-        if LOCAL_CONFIG_PATH.exists():
-            with open(LOCAL_CONFIG_PATH, "r", encoding="utf-8") as f:
-                data = json.load(f)
-
-        data["admin_qqs"] = admin_qqs
-
-        with open(LOCAL_CONFIG_PATH, "w", encoding="utf-8") as f:
-            json.dump(data, f, ensure_ascii=False, indent=2)
-
-        logger.info("已保存管理员列表到 %s", LOCAL_CONFIG_PATH)
-    except Exception as exc:
-        logger.error("保存本地配置失败: %s", exc)
-        raise
-
-
-@dataclass
-class WebUISettings:
-    url: str
-    port: int
-    password: str
-    using_default_password: bool
-    config_exists: bool
-
-    @property
-    def display_url(self) -> str:
-        """用于日志和展示的格式化 URL。"""
-        from Undefined.config.models import format_netloc
-
-        return f"http://{format_netloc(self.url or '0.0.0.0', self.port)}"
-
-
-def load_webui_settings(config_path: Optional[Path] = None) -> WebUISettings:
-    data = load_toml_data(config_path)
-    config_exists = bool(data)
-    url_value = _get_value(data, ("webui", "url"), None)
-    port_value = _get_value(data, ("webui", "port"), None)
-    password_value = _get_value(data, ("webui", "password"), None)
-
-    url = _coerce_str(url_value, DEFAULT_WEBUI_URL)
-    port = _coerce_int(port_value, DEFAULT_WEBUI_PORT)
-    if port <= 0 or port > 65535:
-        port = DEFAULT_WEBUI_PORT
-
-    password_normalized = _normalize_str(password_value)
-    if not password_normalized:
-        return WebUISettings(
-            url=url,
-            port=port,
-            password=DEFAULT_WEBUI_PASSWORD,
-            using_default_password=True,
-            config_exists=config_exists,
-        )
-    return WebUISettings(
-        url=url,
-        port=port,
-        password=password_normalized,
-        using_default_password=False,
-        config_exists=config_exists,
-    )
 
 
 @dataclass
@@ -783,8 +532,8 @@ class Config:
             _get_value(data, ("onebot", "token"), "ONEBOT_TOKEN"), ""
         )
 
-        embedding_model = cls._parse_embedding_model_config(data)
-        rerank_model = cls._parse_rerank_model_config(data)
+        embedding_model = _parse_embedding_model_config(data)
+        rerank_model = _parse_rerank_model_config(data)
 
         knowledge_enabled = _coerce_bool(
             _get_value(data, ("knowledge", "enabled"), None), False
@@ -858,8 +607,8 @@ class Config:
                 )
             knowledge_rerank_top_k = fallback
 
-        chat_model = cls._parse_chat_model_config(data)
-        vision_model = cls._parse_vision_model_config(data)
+        chat_model = _parse_chat_model_config(data)
+        vision_model = _parse_vision_model_config(data)
         security_model_enabled = _coerce_bool(
             _get_value(
                 data,
@@ -868,20 +617,20 @@ class Config:
             ),
             True,
         )
-        security_model = cls._parse_security_model_config(data, chat_model)
-        naga_model = cls._parse_naga_model_config(data, security_model)
-        agent_model = cls._parse_agent_model_config(data)
-        historian_model = cls._parse_historian_model_config(data, agent_model)
-        summary_model, summary_model_configured = cls._parse_summary_model_config(
+        security_model = _parse_security_model_config(data, chat_model)
+        naga_model = _parse_naga_model_config(data, security_model)
+        agent_model = _parse_agent_model_config(data)
+        historian_model = _parse_historian_model_config(data, agent_model)
+        summary_model, summary_model_configured = _parse_summary_model_config(
             data, agent_model
         )
-        grok_model = cls._parse_grok_model_config(data)
+        grok_model = _parse_grok_model_config(data)
 
         model_pool_enabled = _coerce_bool(
             _get_value(data, ("features", "pool_enabled"), "MODEL_POOL_ENABLED"), False
         )
 
-        superadmin_qq, admin_qqs = cls._merge_admins(
+        superadmin_qq, admin_qqs = _merge_admins(
             superadmin_qq=superadmin_qq, admin_qqs=admin_qqs
         )
 
@@ -1013,7 +762,7 @@ class Config:
                 if easter_egg_mode_raw is not None:
                     _warn_env_fallback("EASTER_EGG_CALL_MESSAGE_MODE")
 
-        easter_egg_agent_call_message_mode = cls._parse_easter_egg_call_mode(
+        easter_egg_agent_call_message_mode = _parse_easter_egg_call_mode(
             easter_egg_mode_raw
         )
 
@@ -1443,17 +1192,17 @@ class Config:
             messages_send_url_file_max_size_mb = 100
 
         webui_settings = load_webui_settings(config_path)
-        api_config = cls._parse_api_config(data)
+        api_config = _parse_api_config(data)
 
-        cognitive = cls._parse_cognitive_config(data)
-        memes = cls._parse_memes_config(data)
-        naga = cls._parse_naga_config(data)
-        models_image_gen = cls._parse_image_gen_model_config(data)
-        models_image_edit = cls._parse_image_edit_model_config(data)
-        image_gen = cls._parse_image_gen_config(data)
+        cognitive = _parse_cognitive_config(data)
+        memes = _parse_memes_config(data)
+        naga = _parse_naga_config(data)
+        models_image_gen = _parse_image_gen_model_config(data)
+        models_image_edit = _parse_image_edit_model_config(data)
+        image_gen = _parse_image_gen_config(data)
 
         if strict:
-            cls._verify_required_fields(
+            _verify_required_fields(
                 bot_qq=bot_qq,
                 superadmin_qq=superadmin_qq,
                 onebot_ws_url=onebot_ws_url,
@@ -1464,7 +1213,7 @@ class Config:
                 embedding_model=embedding_model,
             )
 
-        cls._log_debug_info(
+        _log_debug_info(
             chat_model,
             vision_model,
             security_model,
@@ -1793,1056 +1542,6 @@ class Config:
 
         return bool(self.security_model_enabled)
 
-    @staticmethod
-    def _parse_model_pool(
-        data: dict[str, Any],
-        model_section: str,
-        primary_config: ChatModelConfig | AgentModelConfig,
-    ) -> ModelPool | None:
-        """解析模型池配置，缺省字段继承 primary_config"""
-        pool_data = data.get("models", {}).get(model_section, {}).get("pool")
-        if not isinstance(pool_data, dict):
-            return None
-
-        enabled = _coerce_bool(pool_data.get("enabled"), False)
-        strategy = _coerce_str(pool_data.get("strategy"), "default").strip().lower()
-        if strategy not in ("default", "round_robin", "random"):
-            strategy = "default"
-
-        raw_models = pool_data.get("models")
-        if not isinstance(raw_models, list):
-            return ModelPool(enabled=enabled, strategy=strategy)
-
-        entries: list[ModelPoolEntry] = []
-        for item in raw_models:
-            if not isinstance(item, dict):
-                continue
-            name = _coerce_str(item.get("model_name"), "").strip()
-            if not name:
-                continue
-            entries.append(
-                ModelPoolEntry(
-                    api_url=_coerce_str(item.get("api_url"), primary_config.api_url),
-                    api_key=_coerce_str(item.get("api_key"), primary_config.api_key),
-                    model_name=name,
-                    max_tokens=_coerce_int(
-                        item.get("max_tokens"), primary_config.max_tokens
-                    ),
-                    queue_interval_seconds=_normalize_queue_interval(
-                        _coerce_float(
-                            item.get("queue_interval_seconds"),
-                            primary_config.queue_interval_seconds,
-                        ),
-                        primary_config.queue_interval_seconds,
-                    ),
-                    api_mode=(
-                        _coerce_str(item.get("api_mode"), primary_config.api_mode)
-                        .strip()
-                        .lower()
-                    )
-                    if _coerce_str(item.get("api_mode"), primary_config.api_mode)
-                    .strip()
-                    .lower()
-                    in _VALID_API_MODES
-                    else primary_config.api_mode,
-                    thinking_enabled=_coerce_bool(
-                        item.get("thinking_enabled"), primary_config.thinking_enabled
-                    ),
-                    thinking_budget_tokens=_coerce_int(
-                        item.get("thinking_budget_tokens"),
-                        primary_config.thinking_budget_tokens,
-                    ),
-                    thinking_include_budget=_coerce_bool(
-                        item.get("thinking_include_budget"),
-                        primary_config.thinking_include_budget,
-                    ),
-                    reasoning_effort_style=_resolve_reasoning_effort_style(
-                        item.get("reasoning_effort_style"),
-                        primary_config.reasoning_effort_style,
-                    ),
-                    thinking_tool_call_compat=_coerce_bool(
-                        item.get("thinking_tool_call_compat"),
-                        primary_config.thinking_tool_call_compat,
-                    ),
-                    responses_tool_choice_compat=_coerce_bool(
-                        item.get("responses_tool_choice_compat"),
-                        primary_config.responses_tool_choice_compat,
-                    ),
-                    responses_force_stateless_replay=_coerce_bool(
-                        item.get("responses_force_stateless_replay"),
-                        primary_config.responses_force_stateless_replay,
-                    ),
-                    prompt_cache_enabled=_coerce_bool(
-                        item.get("prompt_cache_enabled"),
-                        primary_config.prompt_cache_enabled,
-                    ),
-                    reasoning_enabled=_coerce_bool(
-                        item.get("reasoning_enabled"),
-                        primary_config.reasoning_enabled,
-                    ),
-                    reasoning_effort=_resolve_reasoning_effort(
-                        item.get("reasoning_effort"),
-                        primary_config.reasoning_effort,
-                    ),
-                    request_params=merge_request_params(
-                        primary_config.request_params,
-                        item.get("request_params"),
-                    ),
-                )
-            )
-
-        return ModelPool(enabled=enabled, strategy=strategy, models=entries)
-
-    @staticmethod
-    def _parse_embedding_model_config(data: dict[str, Any]) -> EmbeddingModelConfig:
-        return EmbeddingModelConfig(
-            api_url=_coerce_str(
-                _get_value(
-                    data, ("models", "embedding", "api_url"), "EMBEDDING_MODEL_API_URL"
-                ),
-                "",
-            ),
-            api_key=_coerce_str(
-                _get_value(
-                    data, ("models", "embedding", "api_key"), "EMBEDDING_MODEL_API_KEY"
-                ),
-                "",
-            ),
-            model_name=_coerce_str(
-                _get_value(
-                    data, ("models", "embedding", "model_name"), "EMBEDDING_MODEL_NAME"
-                ),
-                "",
-            ),
-            queue_interval_seconds=_normalize_queue_interval(
-                _coerce_float(
-                    _get_value(
-                        data, ("models", "embedding", "queue_interval_seconds"), None
-                    ),
-                    0.0,
-                ),
-                0.0,
-            ),
-            dimensions=_coerce_int(
-                _get_value(data, ("models", "embedding", "dimensions"), None), 0
-            )
-            or None,
-            query_instruction=_coerce_str(
-                _get_value(data, ("models", "embedding", "query_instruction"), None), ""
-            ),
-            document_instruction=_coerce_str(
-                _get_value(data, ("models", "embedding", "document_instruction"), None),
-                "",
-            ),
-            request_params=_get_model_request_params(data, "embedding"),
-        )
-
-    @staticmethod
-    def _parse_rerank_model_config(data: dict[str, Any]) -> RerankModelConfig:
-        queue_interval_seconds = _normalize_queue_interval(
-            _coerce_float(
-                _get_value(data, ("models", "rerank", "queue_interval_seconds"), None),
-                0.0,
-            ),
-            0.0,
-        )
-        return RerankModelConfig(
-            api_url=_coerce_str(
-                _get_value(
-                    data, ("models", "rerank", "api_url"), "RERANK_MODEL_API_URL"
-                ),
-                "",
-            ),
-            api_key=_coerce_str(
-                _get_value(
-                    data, ("models", "rerank", "api_key"), "RERANK_MODEL_API_KEY"
-                ),
-                "",
-            ),
-            model_name=_coerce_str(
-                _get_value(
-                    data, ("models", "rerank", "model_name"), "RERANK_MODEL_NAME"
-                ),
-                "",
-            ),
-            queue_interval_seconds=queue_interval_seconds,
-            query_instruction=_coerce_str(
-                _get_value(data, ("models", "rerank", "query_instruction"), None), ""
-            ),
-            request_params=_get_model_request_params(data, "rerank"),
-        )
-
-    @staticmethod
-    def _parse_chat_model_config(data: dict[str, Any]) -> ChatModelConfig:
-        queue_interval_seconds = _normalize_queue_interval(
-            _coerce_float(
-                _get_value(
-                    data,
-                    ("models", "chat", "queue_interval_seconds"),
-                    "CHAT_MODEL_QUEUE_INTERVAL",
-                ),
-                1.0,
-            )
-        )
-        thinking_include_budget, thinking_tool_call_compat = (
-            _resolve_thinking_compat_flags(
-                data=data,
-                model_name="chat",
-                include_budget_env_key="CHAT_MODEL_THINKING_INCLUDE_BUDGET",
-                tool_call_compat_env_key="CHAT_MODEL_THINKING_TOOL_CALL_COMPAT",
-                legacy_env_key="CHAT_MODEL_DEEPSEEK_NEW_COT_SUPPORT",
-            )
-        )
-        api_mode = _resolve_api_mode(data, "chat", "CHAT_MODEL_API_MODE")
-        responses_tool_choice_compat = _resolve_responses_tool_choice_compat(
-            data, "chat", "CHAT_MODEL_RESPONSES_TOOL_CHOICE_COMPAT"
-        )
-        responses_force_stateless_replay = _resolve_responses_force_stateless_replay(
-            data, "chat", "CHAT_MODEL_RESPONSES_FORCE_STATELESS_REPLAY"
-        )
-        prompt_cache_enabled = _coerce_bool(
-            _get_value(
-                data,
-                ("models", "chat", "prompt_cache_enabled"),
-                "CHAT_MODEL_PROMPT_CACHE_ENABLED",
-            ),
-            True,
-        )
-        reasoning_enabled = _coerce_bool(
-            _get_value(
-                data,
-                ("models", "chat", "reasoning_enabled"),
-                "CHAT_MODEL_REASONING_ENABLED",
-            ),
-            False,
-        )
-        reasoning_effort = _resolve_reasoning_effort(
-            _get_value(
-                data,
-                ("models", "chat", "reasoning_effort"),
-                "CHAT_MODEL_REASONING_EFFORT",
-            ),
-            "medium",
-        )
-        config = ChatModelConfig(
-            api_url=_coerce_str(
-                _get_value(data, ("models", "chat", "api_url"), "CHAT_MODEL_API_URL"),
-                "",
-            ),
-            api_key=_coerce_str(
-                _get_value(data, ("models", "chat", "api_key"), "CHAT_MODEL_API_KEY"),
-                "",
-            ),
-            model_name=_coerce_str(
-                _get_value(data, ("models", "chat", "model_name"), "CHAT_MODEL_NAME"),
-                "",
-            ),
-            max_tokens=_coerce_int(
-                _get_value(
-                    data, ("models", "chat", "max_tokens"), "CHAT_MODEL_MAX_TOKENS"
-                ),
-                8192,
-            ),
-            queue_interval_seconds=queue_interval_seconds,
-            api_mode=api_mode,
-            thinking_enabled=_coerce_bool(
-                _get_value(
-                    data,
-                    ("models", "chat", "thinking_enabled"),
-                    "CHAT_MODEL_THINKING_ENABLED",
-                ),
-                False,
-            ),
-            thinking_budget_tokens=_coerce_int(
-                _get_value(
-                    data,
-                    ("models", "chat", "thinking_budget_tokens"),
-                    "CHAT_MODEL_THINKING_BUDGET_TOKENS",
-                ),
-                20000,
-            ),
-            thinking_include_budget=thinking_include_budget,
-            reasoning_effort_style=_resolve_reasoning_effort_style(
-                _get_value(
-                    data,
-                    ("models", "chat", "reasoning_effort_style"),
-                    "CHAT_MODEL_REASONING_EFFORT_STYLE",
-                ),
-            ),
-            thinking_tool_call_compat=thinking_tool_call_compat,
-            responses_tool_choice_compat=responses_tool_choice_compat,
-            responses_force_stateless_replay=responses_force_stateless_replay,
-            prompt_cache_enabled=prompt_cache_enabled,
-            reasoning_enabled=reasoning_enabled,
-            reasoning_effort=reasoning_effort,
-            request_params=_get_model_request_params(data, "chat"),
-        )
-        config.pool = Config._parse_model_pool(data, "chat", config)
-        return config
-
-    @staticmethod
-    def _parse_vision_model_config(data: dict[str, Any]) -> VisionModelConfig:
-        queue_interval_seconds = _normalize_queue_interval(
-            _coerce_float(
-                _get_value(
-                    data,
-                    ("models", "vision", "queue_interval_seconds"),
-                    "VISION_MODEL_QUEUE_INTERVAL",
-                ),
-                1.0,
-            )
-        )
-        thinking_include_budget, thinking_tool_call_compat = (
-            _resolve_thinking_compat_flags(
-                data=data,
-                model_name="vision",
-                include_budget_env_key="VISION_MODEL_THINKING_INCLUDE_BUDGET",
-                tool_call_compat_env_key="VISION_MODEL_THINKING_TOOL_CALL_COMPAT",
-                legacy_env_key="VISION_MODEL_DEEPSEEK_NEW_COT_SUPPORT",
-            )
-        )
-        api_mode = _resolve_api_mode(data, "vision", "VISION_MODEL_API_MODE")
-        responses_tool_choice_compat = _resolve_responses_tool_choice_compat(
-            data, "vision", "VISION_MODEL_RESPONSES_TOOL_CHOICE_COMPAT"
-        )
-        responses_force_stateless_replay = _resolve_responses_force_stateless_replay(
-            data, "vision", "VISION_MODEL_RESPONSES_FORCE_STATELESS_REPLAY"
-        )
-        prompt_cache_enabled = _coerce_bool(
-            _get_value(
-                data,
-                ("models", "vision", "prompt_cache_enabled"),
-                "VISION_MODEL_PROMPT_CACHE_ENABLED",
-            ),
-            True,
-        )
-        reasoning_enabled = _coerce_bool(
-            _get_value(
-                data,
-                ("models", "vision", "reasoning_enabled"),
-                "VISION_MODEL_REASONING_ENABLED",
-            ),
-            False,
-        )
-        reasoning_effort = _resolve_reasoning_effort(
-            _get_value(
-                data,
-                ("models", "vision", "reasoning_effort"),
-                "VISION_MODEL_REASONING_EFFORT",
-            ),
-            "medium",
-        )
-        return VisionModelConfig(
-            api_url=_coerce_str(
-                _get_value(
-                    data, ("models", "vision", "api_url"), "VISION_MODEL_API_URL"
-                ),
-                "",
-            ),
-            api_key=_coerce_str(
-                _get_value(
-                    data, ("models", "vision", "api_key"), "VISION_MODEL_API_KEY"
-                ),
-                "",
-            ),
-            model_name=_coerce_str(
-                _get_value(
-                    data, ("models", "vision", "model_name"), "VISION_MODEL_NAME"
-                ),
-                "",
-            ),
-            queue_interval_seconds=queue_interval_seconds,
-            api_mode=api_mode,
-            thinking_enabled=_coerce_bool(
-                _get_value(
-                    data,
-                    ("models", "vision", "thinking_enabled"),
-                    "VISION_MODEL_THINKING_ENABLED",
-                ),
-                False,
-            ),
-            thinking_budget_tokens=_coerce_int(
-                _get_value(
-                    data,
-                    ("models", "vision", "thinking_budget_tokens"),
-                    "VISION_MODEL_THINKING_BUDGET_TOKENS",
-                ),
-                20000,
-            ),
-            thinking_include_budget=thinking_include_budget,
-            reasoning_effort_style=_resolve_reasoning_effort_style(
-                _get_value(
-                    data,
-                    ("models", "vision", "reasoning_effort_style"),
-                    "VISION_MODEL_REASONING_EFFORT_STYLE",
-                ),
-            ),
-            thinking_tool_call_compat=thinking_tool_call_compat,
-            responses_tool_choice_compat=responses_tool_choice_compat,
-            responses_force_stateless_replay=responses_force_stateless_replay,
-            prompt_cache_enabled=prompt_cache_enabled,
-            reasoning_enabled=reasoning_enabled,
-            reasoning_effort=reasoning_effort,
-            request_params=_get_model_request_params(data, "vision"),
-        )
-
-    @staticmethod
-    def _parse_security_model_config(
-        data: dict[str, Any], chat_model: ChatModelConfig
-    ) -> SecurityModelConfig:
-        api_url = _coerce_str(
-            _get_value(
-                data, ("models", "security", "api_url"), "SECURITY_MODEL_API_URL"
-            ),
-            "",
-        )
-        api_key = _coerce_str(
-            _get_value(
-                data, ("models", "security", "api_key"), "SECURITY_MODEL_API_KEY"
-            ),
-            "",
-        )
-        model_name = _coerce_str(
-            _get_value(
-                data, ("models", "security", "model_name"), "SECURITY_MODEL_NAME"
-            ),
-            "",
-        )
-        queue_interval_seconds = _coerce_float(
-            _get_value(
-                data,
-                ("models", "security", "queue_interval_seconds"),
-                "SECURITY_MODEL_QUEUE_INTERVAL",
-            ),
-            1.0,
-        )
-        queue_interval_seconds = _normalize_queue_interval(queue_interval_seconds)
-
-        thinking_include_budget, thinking_tool_call_compat = (
-            _resolve_thinking_compat_flags(
-                data=data,
-                model_name="security",
-                include_budget_env_key="SECURITY_MODEL_THINKING_INCLUDE_BUDGET",
-                tool_call_compat_env_key="SECURITY_MODEL_THINKING_TOOL_CALL_COMPAT",
-                legacy_env_key="SECURITY_MODEL_DEEPSEEK_NEW_COT_SUPPORT",
-            )
-        )
-        api_mode = _resolve_api_mode(data, "security", "SECURITY_MODEL_API_MODE")
-        responses_tool_choice_compat = _resolve_responses_tool_choice_compat(
-            data, "security", "SECURITY_MODEL_RESPONSES_TOOL_CHOICE_COMPAT"
-        )
-        responses_force_stateless_replay = _resolve_responses_force_stateless_replay(
-            data, "security", "SECURITY_MODEL_RESPONSES_FORCE_STATELESS_REPLAY"
-        )
-        prompt_cache_enabled = _coerce_bool(
-            _get_value(
-                data,
-                ("models", "security", "prompt_cache_enabled"),
-                "SECURITY_MODEL_PROMPT_CACHE_ENABLED",
-            ),
-            True,
-        )
-        reasoning_enabled = _coerce_bool(
-            _get_value(
-                data,
-                ("models", "security", "reasoning_enabled"),
-                "SECURITY_MODEL_REASONING_ENABLED",
-            ),
-            False,
-        )
-        reasoning_effort = _resolve_reasoning_effort(
-            _get_value(
-                data,
-                ("models", "security", "reasoning_effort"),
-                "SECURITY_MODEL_REASONING_EFFORT",
-            ),
-            "medium",
-        )
-
-        if api_url and api_key and model_name:
-            return SecurityModelConfig(
-                api_url=api_url,
-                api_key=api_key,
-                model_name=model_name,
-                max_tokens=_coerce_int(
-                    _get_value(
-                        data,
-                        ("models", "security", "max_tokens"),
-                        "SECURITY_MODEL_MAX_TOKENS",
-                    ),
-                    100,
-                ),
-                queue_interval_seconds=queue_interval_seconds,
-                api_mode=api_mode,
-                thinking_enabled=_coerce_bool(
-                    _get_value(
-                        data,
-                        ("models", "security", "thinking_enabled"),
-                        "SECURITY_MODEL_THINKING_ENABLED",
-                    ),
-                    False,
-                ),
-                thinking_budget_tokens=_coerce_int(
-                    _get_value(
-                        data,
-                        ("models", "security", "thinking_budget_tokens"),
-                        "SECURITY_MODEL_THINKING_BUDGET_TOKENS",
-                    ),
-                    0,
-                ),
-                thinking_include_budget=thinking_include_budget,
-                reasoning_effort_style=_resolve_reasoning_effort_style(
-                    _get_value(
-                        data,
-                        ("models", "security", "reasoning_effort_style"),
-                        "SECURITY_MODEL_REASONING_EFFORT_STYLE",
-                    ),
-                ),
-                thinking_tool_call_compat=thinking_tool_call_compat,
-                responses_tool_choice_compat=responses_tool_choice_compat,
-                responses_force_stateless_replay=responses_force_stateless_replay,
-                prompt_cache_enabled=prompt_cache_enabled,
-                reasoning_enabled=reasoning_enabled,
-                reasoning_effort=reasoning_effort,
-                request_params=_get_model_request_params(data, "security"),
-            )
-
-        logger.warning("未配置安全模型，将使用对话模型作为后备")
-        return SecurityModelConfig(
-            api_url=chat_model.api_url,
-            api_key=chat_model.api_key,
-            model_name=chat_model.model_name,
-            max_tokens=chat_model.max_tokens,
-            queue_interval_seconds=chat_model.queue_interval_seconds,
-            api_mode=chat_model.api_mode,
-            thinking_enabled=False,
-            thinking_budget_tokens=0,
-            thinking_include_budget=True,
-            reasoning_effort_style="openai",
-            thinking_tool_call_compat=chat_model.thinking_tool_call_compat,
-            responses_tool_choice_compat=chat_model.responses_tool_choice_compat,
-            responses_force_stateless_replay=chat_model.responses_force_stateless_replay,
-            prompt_cache_enabled=chat_model.prompt_cache_enabled,
-            reasoning_enabled=chat_model.reasoning_enabled,
-            reasoning_effort=chat_model.reasoning_effort,
-            request_params=merge_request_params(chat_model.request_params),
-        )
-
-    @staticmethod
-    def _parse_naga_model_config(
-        data: dict[str, Any], security_model: SecurityModelConfig
-    ) -> SecurityModelConfig:
-        api_url = _coerce_str(
-            _get_value(data, ("models", "naga", "api_url"), "NAGA_MODEL_API_URL"),
-            "",
-        )
-        api_key = _coerce_str(
-            _get_value(data, ("models", "naga", "api_key"), "NAGA_MODEL_API_KEY"),
-            "",
-        )
-        model_name = _coerce_str(
-            _get_value(data, ("models", "naga", "model_name"), "NAGA_MODEL_NAME"),
-            "",
-        )
-        queue_interval_seconds = _coerce_float(
-            _get_value(
-                data,
-                ("models", "naga", "queue_interval_seconds"),
-                "NAGA_MODEL_QUEUE_INTERVAL",
-            ),
-            security_model.queue_interval_seconds,
-        )
-        queue_interval_seconds = _normalize_queue_interval(queue_interval_seconds)
-
-        thinking_include_budget, thinking_tool_call_compat = (
-            _resolve_thinking_compat_flags(
-                data=data,
-                model_name="naga",
-                include_budget_env_key="NAGA_MODEL_THINKING_INCLUDE_BUDGET",
-                tool_call_compat_env_key="NAGA_MODEL_THINKING_TOOL_CALL_COMPAT",
-                legacy_env_key="NAGA_MODEL_DEEPSEEK_NEW_COT_SUPPORT",
-            )
-        )
-        api_mode = _resolve_api_mode(data, "naga", "NAGA_MODEL_API_MODE")
-        responses_tool_choice_compat = _resolve_responses_tool_choice_compat(
-            data, "naga", "NAGA_MODEL_RESPONSES_TOOL_CHOICE_COMPAT"
-        )
-        responses_force_stateless_replay = _resolve_responses_force_stateless_replay(
-            data, "naga", "NAGA_MODEL_RESPONSES_FORCE_STATELESS_REPLAY"
-        )
-        prompt_cache_enabled = _coerce_bool(
-            _get_value(
-                data,
-                ("models", "naga", "prompt_cache_enabled"),
-                "NAGA_MODEL_PROMPT_CACHE_ENABLED",
-            ),
-            getattr(security_model, "prompt_cache_enabled", True),
-        )
-        reasoning_enabled = _coerce_bool(
-            _get_value(
-                data,
-                ("models", "naga", "reasoning_enabled"),
-                "NAGA_MODEL_REASONING_ENABLED",
-            ),
-            getattr(security_model, "reasoning_enabled", False),
-        )
-        reasoning_effort = _resolve_reasoning_effort(
-            _get_value(
-                data,
-                ("models", "naga", "reasoning_effort"),
-                "NAGA_MODEL_REASONING_EFFORT",
-            ),
-            getattr(security_model, "reasoning_effort", "medium"),
-        )
-
-        if api_url and api_key and model_name:
-            return SecurityModelConfig(
-                api_url=api_url,
-                api_key=api_key,
-                model_name=model_name,
-                max_tokens=_coerce_int(
-                    _get_value(
-                        data,
-                        ("models", "naga", "max_tokens"),
-                        "NAGA_MODEL_MAX_TOKENS",
-                    ),
-                    160,
-                ),
-                queue_interval_seconds=queue_interval_seconds,
-                api_mode=api_mode,
-                thinking_enabled=_coerce_bool(
-                    _get_value(
-                        data,
-                        ("models", "naga", "thinking_enabled"),
-                        "NAGA_MODEL_THINKING_ENABLED",
-                    ),
-                    False,
-                ),
-                thinking_budget_tokens=_coerce_int(
-                    _get_value(
-                        data,
-                        ("models", "naga", "thinking_budget_tokens"),
-                        "NAGA_MODEL_THINKING_BUDGET_TOKENS",
-                    ),
-                    0,
-                ),
-                thinking_include_budget=thinking_include_budget,
-                reasoning_effort_style=_resolve_reasoning_effort_style(
-                    _get_value(
-                        data,
-                        ("models", "naga", "reasoning_effort_style"),
-                        "NAGA_MODEL_REASONING_EFFORT_STYLE",
-                    ),
-                ),
-                thinking_tool_call_compat=thinking_tool_call_compat,
-                responses_tool_choice_compat=responses_tool_choice_compat,
-                responses_force_stateless_replay=responses_force_stateless_replay,
-                prompt_cache_enabled=prompt_cache_enabled,
-                reasoning_enabled=reasoning_enabled,
-                reasoning_effort=reasoning_effort,
-                request_params=_get_model_request_params(data, "naga"),
-            )
-
-        logger.info(
-            "未配置 Naga 审核模型，将使用已解析的安全模型配置作为后备（安全模型本身可能已回退）"
-        )
-        return SecurityModelConfig(
-            api_url=security_model.api_url,
-            api_key=security_model.api_key,
-            model_name=security_model.model_name,
-            max_tokens=security_model.max_tokens,
-            queue_interval_seconds=security_model.queue_interval_seconds,
-            api_mode=security_model.api_mode,
-            thinking_enabled=security_model.thinking_enabled,
-            thinking_budget_tokens=security_model.thinking_budget_tokens,
-            thinking_include_budget=security_model.thinking_include_budget,
-            reasoning_effort_style=security_model.reasoning_effort_style,
-            thinking_tool_call_compat=security_model.thinking_tool_call_compat,
-            responses_tool_choice_compat=security_model.responses_tool_choice_compat,
-            responses_force_stateless_replay=security_model.responses_force_stateless_replay,
-            prompt_cache_enabled=security_model.prompt_cache_enabled,
-            reasoning_enabled=security_model.reasoning_enabled,
-            reasoning_effort=security_model.reasoning_effort,
-            request_params=merge_request_params(security_model.request_params),
-        )
-
-    @staticmethod
-    def _parse_agent_model_config(data: dict[str, Any]) -> AgentModelConfig:
-        queue_interval_seconds = _normalize_queue_interval(
-            _coerce_float(
-                _get_value(
-                    data,
-                    ("models", "agent", "queue_interval_seconds"),
-                    "AGENT_MODEL_QUEUE_INTERVAL",
-                ),
-                1.0,
-            )
-        )
-        thinking_include_budget, thinking_tool_call_compat = (
-            _resolve_thinking_compat_flags(
-                data=data,
-                model_name="agent",
-                include_budget_env_key="AGENT_MODEL_THINKING_INCLUDE_BUDGET",
-                tool_call_compat_env_key="AGENT_MODEL_THINKING_TOOL_CALL_COMPAT",
-                legacy_env_key="AGENT_MODEL_DEEPSEEK_NEW_COT_SUPPORT",
-            )
-        )
-        api_mode = _resolve_api_mode(data, "agent", "AGENT_MODEL_API_MODE")
-        responses_tool_choice_compat = _resolve_responses_tool_choice_compat(
-            data, "agent", "AGENT_MODEL_RESPONSES_TOOL_CHOICE_COMPAT"
-        )
-        responses_force_stateless_replay = _resolve_responses_force_stateless_replay(
-            data, "agent", "AGENT_MODEL_RESPONSES_FORCE_STATELESS_REPLAY"
-        )
-        prompt_cache_enabled = _coerce_bool(
-            _get_value(
-                data,
-                ("models", "agent", "prompt_cache_enabled"),
-                "AGENT_MODEL_PROMPT_CACHE_ENABLED",
-            ),
-            True,
-        )
-        reasoning_enabled = _coerce_bool(
-            _get_value(
-                data,
-                ("models", "agent", "reasoning_enabled"),
-                "AGENT_MODEL_REASONING_ENABLED",
-            ),
-            False,
-        )
-        reasoning_effort = _resolve_reasoning_effort(
-            _get_value(
-                data,
-                ("models", "agent", "reasoning_effort"),
-                "AGENT_MODEL_REASONING_EFFORT",
-            ),
-            "medium",
-        )
-        config = AgentModelConfig(
-            api_url=_coerce_str(
-                _get_value(data, ("models", "agent", "api_url"), "AGENT_MODEL_API_URL"),
-                "",
-            ),
-            api_key=_coerce_str(
-                _get_value(data, ("models", "agent", "api_key"), "AGENT_MODEL_API_KEY"),
-                "",
-            ),
-            model_name=_coerce_str(
-                _get_value(data, ("models", "agent", "model_name"), "AGENT_MODEL_NAME"),
-                "",
-            ),
-            max_tokens=_coerce_int(
-                _get_value(
-                    data, ("models", "agent", "max_tokens"), "AGENT_MODEL_MAX_TOKENS"
-                ),
-                4096,
-            ),
-            queue_interval_seconds=queue_interval_seconds,
-            api_mode=api_mode,
-            thinking_enabled=_coerce_bool(
-                _get_value(
-                    data,
-                    ("models", "agent", "thinking_enabled"),
-                    "AGENT_MODEL_THINKING_ENABLED",
-                ),
-                False,
-            ),
-            thinking_budget_tokens=_coerce_int(
-                _get_value(
-                    data,
-                    ("models", "agent", "thinking_budget_tokens"),
-                    "AGENT_MODEL_THINKING_BUDGET_TOKENS",
-                ),
-                0,
-            ),
-            thinking_include_budget=thinking_include_budget,
-            reasoning_effort_style=_resolve_reasoning_effort_style(
-                _get_value(
-                    data,
-                    ("models", "agent", "reasoning_effort_style"),
-                    "AGENT_MODEL_REASONING_EFFORT_STYLE",
-                ),
-            ),
-            thinking_tool_call_compat=thinking_tool_call_compat,
-            responses_tool_choice_compat=responses_tool_choice_compat,
-            responses_force_stateless_replay=responses_force_stateless_replay,
-            prompt_cache_enabled=prompt_cache_enabled,
-            reasoning_enabled=reasoning_enabled,
-            reasoning_effort=reasoning_effort,
-            request_params=_get_model_request_params(data, "agent"),
-        )
-        config.pool = Config._parse_model_pool(data, "agent", config)
-        return config
-
-    @staticmethod
-    def _parse_grok_model_config(data: dict[str, Any]) -> GrokModelConfig:
-        queue_interval_seconds = _normalize_queue_interval(
-            _coerce_float(
-                _get_value(
-                    data,
-                    ("models", "grok", "queue_interval_seconds"),
-                    "GROK_MODEL_QUEUE_INTERVAL",
-                ),
-                1.0,
-            )
-        )
-        return GrokModelConfig(
-            api_url=_coerce_str(
-                _get_value(data, ("models", "grok", "api_url"), "GROK_MODEL_API_URL"),
-                "",
-            ),
-            api_key=_coerce_str(
-                _get_value(data, ("models", "grok", "api_key"), "GROK_MODEL_API_KEY"),
-                "",
-            ),
-            model_name=_coerce_str(
-                _get_value(data, ("models", "grok", "model_name"), "GROK_MODEL_NAME"),
-                "",
-            ),
-            max_tokens=_coerce_int(
-                _get_value(
-                    data, ("models", "grok", "max_tokens"), "GROK_MODEL_MAX_TOKENS"
-                ),
-                8192,
-            ),
-            queue_interval_seconds=queue_interval_seconds,
-            thinking_enabled=_coerce_bool(
-                _get_value(
-                    data,
-                    ("models", "grok", "thinking_enabled"),
-                    "GROK_MODEL_THINKING_ENABLED",
-                ),
-                False,
-            ),
-            thinking_budget_tokens=_coerce_int(
-                _get_value(
-                    data,
-                    ("models", "grok", "thinking_budget_tokens"),
-                    "GROK_MODEL_THINKING_BUDGET_TOKENS",
-                ),
-                20000,
-            ),
-            thinking_include_budget=_coerce_bool(
-                _get_value(
-                    data,
-                    ("models", "grok", "thinking_include_budget"),
-                    "GROK_MODEL_THINKING_INCLUDE_BUDGET",
-                ),
-                True,
-            ),
-            reasoning_effort_style=_resolve_reasoning_effort_style(
-                _get_value(
-                    data,
-                    ("models", "grok", "reasoning_effort_style"),
-                    "GROK_MODEL_REASONING_EFFORT_STYLE",
-                ),
-            ),
-            prompt_cache_enabled=_coerce_bool(
-                _get_value(
-                    data,
-                    ("models", "grok", "prompt_cache_enabled"),
-                    "GROK_MODEL_PROMPT_CACHE_ENABLED",
-                ),
-                True,
-            ),
-            reasoning_enabled=_coerce_bool(
-                _get_value(
-                    data,
-                    ("models", "grok", "reasoning_enabled"),
-                    "GROK_MODEL_REASONING_ENABLED",
-                ),
-                False,
-            ),
-            reasoning_effort=_resolve_reasoning_effort(
-                _get_value(
-                    data,
-                    ("models", "grok", "reasoning_effort"),
-                    "GROK_MODEL_REASONING_EFFORT",
-                ),
-                "medium",
-            ),
-            request_params=_get_model_request_params(data, "grok"),
-        )
-
-    @staticmethod
-    def _parse_image_gen_model_config(data: dict[str, Any]) -> ImageGenModelConfig:
-        """解析 [models.image_gen] 生图模型配置"""
-        return ImageGenModelConfig(
-            api_url=_coerce_str(
-                _get_value(
-                    data, ("models", "image_gen", "api_url"), "IMAGE_GEN_MODEL_API_URL"
-                ),
-                "",
-            ),
-            api_key=_coerce_str(
-                _get_value(
-                    data, ("models", "image_gen", "api_key"), "IMAGE_GEN_MODEL_API_KEY"
-                ),
-                "",
-            ),
-            model_name=_coerce_str(
-                _get_value(
-                    data, ("models", "image_gen", "model_name"), "IMAGE_GEN_MODEL_NAME"
-                ),
-                "",
-            ),
-            request_params=_get_model_request_params(data, "image_gen"),
-        )
-
-    @staticmethod
-    def _parse_image_edit_model_config(data: dict[str, Any]) -> ImageGenModelConfig:
-        """解析 [models.image_edit] 参考图生图模型配置"""
-        return ImageGenModelConfig(
-            api_url=_coerce_str(
-                _get_value(
-                    data,
-                    ("models", "image_edit", "api_url"),
-                    "IMAGE_EDIT_MODEL_API_URL",
-                ),
-                "",
-            ),
-            api_key=_coerce_str(
-                _get_value(
-                    data,
-                    ("models", "image_edit", "api_key"),
-                    "IMAGE_EDIT_MODEL_API_KEY",
-                ),
-                "",
-            ),
-            model_name=_coerce_str(
-                _get_value(
-                    data,
-                    ("models", "image_edit", "model_name"),
-                    "IMAGE_EDIT_MODEL_NAME",
-                ),
-                "",
-            ),
-            request_params=_get_model_request_params(data, "image_edit"),
-        )
-
-    @staticmethod
-    def _parse_image_gen_config(data: dict[str, Any]) -> ImageGenConfig:
-        """解析 [image_gen] 生图工具配置"""
-        return ImageGenConfig(
-            provider=_coerce_str(
-                _get_value(data, ("image_gen", "provider"), "IMAGE_GEN_PROVIDER"),
-                "xingzhige",
-            ),
-            xingzhige_size=_coerce_str(
-                _get_value(data, ("image_gen", "xingzhige_size"), None), "1:1"
-            ),
-            openai_size=_coerce_str(
-                _get_value(data, ("image_gen", "openai_size"), None), ""
-            ),
-            openai_quality=_coerce_str(
-                _get_value(data, ("image_gen", "openai_quality"), None), ""
-            ),
-            openai_style=_coerce_str(
-                _get_value(data, ("image_gen", "openai_style"), None), ""
-            ),
-            openai_timeout=_coerce_float(
-                _get_value(data, ("image_gen", "openai_timeout"), None), 120.0
-            ),
-        )
-
-    @staticmethod
-    def _merge_admins(
-        superadmin_qq: int, admin_qqs: list[int]
-    ) -> tuple[int, list[int]]:
-        local_admins = load_local_admins()
-        all_admins = list(set(admin_qqs + local_admins))
-        if superadmin_qq and superadmin_qq not in all_admins:
-            all_admins.append(superadmin_qq)
-        return superadmin_qq, all_admins
-
-    @staticmethod
-    def _verify_required_fields(
-        bot_qq: int,
-        superadmin_qq: int,
-        onebot_ws_url: str,
-        chat_model: ChatModelConfig,
-        vision_model: VisionModelConfig,
-        agent_model: AgentModelConfig,
-        knowledge_enabled: bool,
-        embedding_model: EmbeddingModelConfig,
-    ) -> None:
-        missing: list[str] = []
-        if bot_qq <= 0:
-            missing.append("core.bot_qq")
-        if superadmin_qq <= 0:
-            missing.append("core.superadmin_qq")
-        if not onebot_ws_url:
-            missing.append("onebot.ws_url")
-        if not chat_model.api_url:
-            missing.append("models.chat.api_url")
-        if not chat_model.api_key:
-            missing.append("models.chat.api_key")
-        if not chat_model.model_name:
-            missing.append("models.chat.model_name")
-        if not vision_model.api_url:
-            missing.append("models.vision.api_url")
-        if not vision_model.api_key:
-            missing.append("models.vision.api_key")
-        if not vision_model.model_name:
-            missing.append("models.vision.model_name")
-        if not agent_model.api_url:
-            missing.append("models.agent.api_url")
-        if not agent_model.api_key:
-            missing.append("models.agent.api_key")
-        if not agent_model.model_name:
-            missing.append("models.agent.model_name")
-        if knowledge_enabled:
-            if not embedding_model.api_url:
-                missing.append("models.embedding.api_url")
-            if not embedding_model.model_name:
-                missing.append("models.embedding.model_name")
-        if missing:
-            raise ValueError(f"缺少必需配置: {', '.join(missing)}")
-
-    @staticmethod
-    def _log_debug_info(
-        chat_model: ChatModelConfig,
-        vision_model: VisionModelConfig,
-        security_model: SecurityModelConfig,
-        naga_model: SecurityModelConfig,
-        agent_model: AgentModelConfig,
-        summary_model: AgentModelConfig,
-        grok_model: GrokModelConfig,
-    ) -> None:
-        configs: list[
-            tuple[
-                str,
-                ChatModelConfig
-                | VisionModelConfig
-                | SecurityModelConfig
-                | AgentModelConfig
-                | GrokModelConfig,
-            ]
-        ] = [
-            ("chat", chat_model),
-            ("vision", vision_model),
-            ("security", security_model),
-            ("naga", naga_model),
-            ("agent", agent_model),
-            ("summary", summary_model),
-            ("grok", grok_model),
-        ]
-        for name, cfg in configs:
-            logger.debug(
-                "[配置] %s_model=%s api_url=%s api_key_set=%s api_mode=%s thinking=%s reasoning=%s/%s cot_compat=%s responses_tool_choice_compat=%s responses_force_stateless_replay=%s",
-                name,
-                cfg.model_name,
-                cfg.api_url,
-                bool(cfg.api_key),
-                getattr(cfg, "api_mode", "chat_completions"),
-                cfg.thinking_enabled,
-                getattr(cfg, "reasoning_enabled", False),
-                getattr(cfg, "reasoning_effort", "medium"),
-                getattr(cfg, "thinking_tool_call_compat", False),
-                getattr(cfg, "responses_tool_choice_compat", False),
-                getattr(cfg, "responses_force_stateless_replay", False),
-            )
-
     def update_from(self, new_config: "Config") -> dict[str, tuple[Any, Any]]:
         changes: dict[str, tuple[Any, Any]] = {}
         for field in fields(self):
@@ -2865,457 +1564,6 @@ class Config:
                 setattr(self, name, new_value)
                 changes[name] = (old_value, new_value)
         return changes
-
-    @staticmethod
-    def _parse_historian_model_config(
-        data: dict[str, Any], fallback: AgentModelConfig
-    ) -> AgentModelConfig:
-        h = data.get("models", {}).get("historian", {})
-        if not isinstance(h, dict) or not h:
-            return fallback
-        queue_interval_seconds = _coerce_float(
-            h.get("queue_interval_seconds"), fallback.queue_interval_seconds
-        )
-        queue_interval_seconds = _normalize_queue_interval(
-            queue_interval_seconds, fallback.queue_interval_seconds
-        )
-        thinking_include_budget, thinking_tool_call_compat = (
-            _resolve_thinking_compat_flags(
-                data={"models": {"historian": h}},
-                model_name="historian",
-                include_budget_env_key="HISTORIAN_MODEL_THINKING_INCLUDE_BUDGET",
-                tool_call_compat_env_key="HISTORIAN_MODEL_THINKING_TOOL_CALL_COMPAT",
-                legacy_env_key="HISTORIAN_MODEL_DEEPSEEK_NEW_COT_SUPPORT",
-            )
-        )
-        api_mode = _resolve_api_mode(
-            {"models": {"historian": h}},
-            "historian",
-            "HISTORIAN_MODEL_API_MODE",
-            fallback.api_mode,
-        )
-        responses_tool_choice_compat = _resolve_responses_tool_choice_compat(
-            {"models": {"historian": h}},
-            "historian",
-            "HISTORIAN_MODEL_RESPONSES_TOOL_CHOICE_COMPAT",
-            fallback.responses_tool_choice_compat,
-        )
-        responses_force_stateless_replay = _resolve_responses_force_stateless_replay(
-            {"models": {"historian": h}},
-            "historian",
-            "HISTORIAN_MODEL_RESPONSES_FORCE_STATELESS_REPLAY",
-            fallback.responses_force_stateless_replay,
-        )
-        prompt_cache_enabled = _coerce_bool(
-            _get_value(
-                {"models": {"historian": h}},
-                ("models", "historian", "prompt_cache_enabled"),
-                "HISTORIAN_MODEL_PROMPT_CACHE_ENABLED",
-            ),
-            fallback.prompt_cache_enabled,
-        )
-        return AgentModelConfig(
-            api_url=_coerce_str(h.get("api_url"), fallback.api_url),
-            api_key=_coerce_str(h.get("api_key"), fallback.api_key),
-            model_name=_coerce_str(h.get("model_name"), fallback.model_name),
-            max_tokens=_coerce_int(h.get("max_tokens"), fallback.max_tokens),
-            queue_interval_seconds=queue_interval_seconds,
-            api_mode=api_mode,
-            thinking_enabled=_coerce_bool(
-                h.get("thinking_enabled"), fallback.thinking_enabled
-            ),
-            thinking_budget_tokens=_coerce_int(
-                h.get("thinking_budget_tokens"), fallback.thinking_budget_tokens
-            ),
-            thinking_include_budget=thinking_include_budget,
-            reasoning_effort_style=_resolve_reasoning_effort_style(
-                _get_value(
-                    {"models": {"historian": h}},
-                    ("models", "historian", "reasoning_effort_style"),
-                    "HISTORIAN_MODEL_REASONING_EFFORT_STYLE",
-                ),
-                fallback.reasoning_effort_style,
-            ),
-            thinking_tool_call_compat=thinking_tool_call_compat,
-            responses_tool_choice_compat=responses_tool_choice_compat,
-            responses_force_stateless_replay=responses_force_stateless_replay,
-            prompt_cache_enabled=prompt_cache_enabled,
-            reasoning_enabled=_coerce_bool(
-                _get_value(
-                    {"models": {"historian": h}},
-                    ("models", "historian", "reasoning_enabled"),
-                    "HISTORIAN_MODEL_REASONING_ENABLED",
-                ),
-                fallback.reasoning_enabled,
-            ),
-            reasoning_effort=_resolve_reasoning_effort(
-                _get_value(
-                    {"models": {"historian": h}},
-                    ("models", "historian", "reasoning_effort"),
-                    "HISTORIAN_MODEL_REASONING_EFFORT",
-                ),
-                fallback.reasoning_effort,
-            ),
-            request_params=merge_request_params(
-                fallback.request_params,
-                h.get("request_params"),
-            ),
-        )
-
-    @staticmethod
-    def _parse_summary_model_config(
-        data: dict[str, Any], fallback: AgentModelConfig
-    ) -> tuple[AgentModelConfig, bool]:
-        s = data.get("models", {}).get("summary", {})
-        if not isinstance(s, dict) or not s:
-            return fallback, False
-        queue_interval_seconds = _coerce_float(
-            s.get("queue_interval_seconds"), fallback.queue_interval_seconds
-        )
-        queue_interval_seconds = _normalize_queue_interval(
-            queue_interval_seconds, fallback.queue_interval_seconds
-        )
-        thinking_include_budget, thinking_tool_call_compat = (
-            _resolve_thinking_compat_flags(
-                data={"models": {"summary": s}},
-                model_name="summary",
-                include_budget_env_key="SUMMARY_MODEL_THINKING_INCLUDE_BUDGET",
-                tool_call_compat_env_key="SUMMARY_MODEL_THINKING_TOOL_CALL_COMPAT",
-                legacy_env_key="SUMMARY_MODEL_DEEPSEEK_NEW_COT_SUPPORT",
-            )
-        )
-        api_mode = _resolve_api_mode(
-            {"models": {"summary": s}},
-            "summary",
-            "SUMMARY_MODEL_API_MODE",
-            fallback.api_mode,
-        )
-        responses_tool_choice_compat = _resolve_responses_tool_choice_compat(
-            {"models": {"summary": s}},
-            "summary",
-            "SUMMARY_MODEL_RESPONSES_TOOL_CHOICE_COMPAT",
-            fallback.responses_tool_choice_compat,
-        )
-        responses_force_stateless_replay = _resolve_responses_force_stateless_replay(
-            {"models": {"summary": s}},
-            "summary",
-            "SUMMARY_MODEL_RESPONSES_FORCE_STATELESS_REPLAY",
-            fallback.responses_force_stateless_replay,
-        )
-        prompt_cache_enabled = _coerce_bool(
-            _get_value(
-                {"models": {"summary": s}},
-                ("models", "summary", "prompt_cache_enabled"),
-                "SUMMARY_MODEL_PROMPT_CACHE_ENABLED",
-            ),
-            fallback.prompt_cache_enabled,
-        )
-        return (
-            AgentModelConfig(
-                api_url=_coerce_str(s.get("api_url"), fallback.api_url),
-                api_key=_coerce_str(s.get("api_key"), fallback.api_key),
-                model_name=_coerce_str(s.get("model_name"), fallback.model_name),
-                max_tokens=_coerce_int(s.get("max_tokens"), fallback.max_tokens),
-                queue_interval_seconds=queue_interval_seconds,
-                api_mode=api_mode,
-                thinking_enabled=_coerce_bool(
-                    s.get("thinking_enabled"), fallback.thinking_enabled
-                ),
-                thinking_budget_tokens=_coerce_int(
-                    s.get("thinking_budget_tokens"), fallback.thinking_budget_tokens
-                ),
-                thinking_include_budget=thinking_include_budget,
-                reasoning_effort_style=_resolve_reasoning_effort_style(
-                    _get_value(
-                        {"models": {"summary": s}},
-                        ("models", "summary", "reasoning_effort_style"),
-                        "SUMMARY_MODEL_REASONING_EFFORT_STYLE",
-                    ),
-                    fallback.reasoning_effort_style,
-                ),
-                thinking_tool_call_compat=thinking_tool_call_compat,
-                responses_tool_choice_compat=responses_tool_choice_compat,
-                responses_force_stateless_replay=responses_force_stateless_replay,
-                prompt_cache_enabled=prompt_cache_enabled,
-                reasoning_enabled=_coerce_bool(
-                    _get_value(
-                        {"models": {"summary": s}},
-                        ("models", "summary", "reasoning_enabled"),
-                        "SUMMARY_MODEL_REASONING_ENABLED",
-                    ),
-                    fallback.reasoning_enabled,
-                ),
-                reasoning_effort=_resolve_reasoning_effort(
-                    _get_value(
-                        {"models": {"summary": s}},
-                        ("models", "summary", "reasoning_effort"),
-                        "SUMMARY_MODEL_REASONING_EFFORT",
-                    ),
-                    fallback.reasoning_effort,
-                ),
-                request_params=merge_request_params(
-                    fallback.request_params,
-                    s.get("request_params"),
-                ),
-            ),
-            True,
-        )
-
-    @staticmethod
-    def _parse_cognitive_config(data: dict[str, Any]) -> CognitiveConfig:
-        cog = data.get("cognitive", {})
-        vs = cog.get("vector_store", {}) if isinstance(cog, dict) else {}
-        q = cog.get("query", {}) if isinstance(cog, dict) else {}
-        hist = cog.get("historian", {}) if isinstance(cog, dict) else {}
-        prof = cog.get("profile", {}) if isinstance(cog, dict) else {}
-        que = cog.get("queue", {}) if isinstance(cog, dict) else {}
-        return CognitiveConfig(
-            enabled=_coerce_bool(
-                cog.get("enabled") if isinstance(cog, dict) else None, True
-            ),
-            bot_name=_coerce_str(
-                cog.get("bot_name") if isinstance(cog, dict) else None,
-                "Undefined",
-            ),
-            vector_store_path=_coerce_str(
-                vs.get("path") if isinstance(vs, dict) else None,
-                "data/cognitive/chromadb",
-            ),
-            queue_path=_coerce_str(
-                que.get("path") if isinstance(que, dict) else None,
-                "data/cognitive/queues",
-            ),
-            profiles_path=_coerce_str(
-                prof.get("path") if isinstance(prof, dict) else None,
-                "data/cognitive/profiles",
-            ),
-            auto_top_k=_coerce_int(
-                q.get("auto_top_k") if isinstance(q, dict) else None, 3
-            ),
-            auto_scope_candidate_multiplier=_coerce_int(
-                q.get("auto_scope_candidate_multiplier")
-                if isinstance(q, dict)
-                else None,
-                2,
-            ),
-            auto_current_group_boost=_coerce_float(
-                q.get("auto_current_group_boost") if isinstance(q, dict) else None,
-                1.15,
-            ),
-            auto_current_private_boost=_coerce_float(
-                q.get("auto_current_private_boost") if isinstance(q, dict) else None,
-                1.25,
-            ),
-            enable_rerank=_coerce_bool(
-                q.get("enable_rerank") if isinstance(q, dict) else None, True
-            ),
-            recent_end_summaries_inject_k=_coerce_int(
-                q.get("recent_end_summaries_inject_k") if isinstance(q, dict) else None,
-                30,
-            ),
-            time_decay_enabled=_coerce_bool(
-                q.get("time_decay_enabled") if isinstance(q, dict) else None, True
-            ),
-            time_decay_half_life_days_auto=_coerce_float(
-                q.get("time_decay_half_life_days_auto")
-                if isinstance(q, dict)
-                else None,
-                14.0,
-            ),
-            time_decay_half_life_days_tool=_coerce_float(
-                q.get("time_decay_half_life_days_tool")
-                if isinstance(q, dict)
-                else None,
-                60.0,
-            ),
-            time_decay_boost=_coerce_float(
-                q.get("time_decay_boost") if isinstance(q, dict) else None, 0.2
-            ),
-            time_decay_min_similarity=_coerce_float(
-                q.get("time_decay_min_similarity") if isinstance(q, dict) else None,
-                0.35,
-            ),
-            tool_default_top_k=_coerce_int(
-                q.get("tool_default_top_k") if isinstance(q, dict) else None, 12
-            ),
-            profile_top_k=_coerce_int(
-                q.get("profile_top_k") if isinstance(q, dict) else None, 8
-            ),
-            rerank_candidate_multiplier=_coerce_int(
-                q.get("rerank_candidate_multiplier") if isinstance(q, dict) else None, 3
-            ),
-            rewrite_max_retry=_coerce_int(
-                hist.get("rewrite_max_retry") if isinstance(hist, dict) else None, 2
-            ),
-            historian_recent_messages_inject_k=_coerce_int(
-                hist.get("recent_messages_inject_k")
-                if isinstance(hist, dict)
-                else None,
-                12,
-            ),
-            historian_recent_message_line_max_len=_coerce_int(
-                hist.get("recent_message_line_max_len")
-                if isinstance(hist, dict)
-                else None,
-                240,
-            ),
-            historian_source_message_max_len=_coerce_int(
-                hist.get("source_message_max_len") if isinstance(hist, dict) else None,
-                800,
-            ),
-            poll_interval_seconds=_coerce_float(
-                hist.get("poll_interval_seconds") if isinstance(hist, dict) else None,
-                1.0,
-            ),
-            stale_job_timeout_seconds=_coerce_float(
-                hist.get("stale_job_timeout_seconds")
-                if isinstance(hist, dict)
-                else None,
-                300.0,
-            ),
-            profile_revision_keep=_coerce_int(
-                prof.get("revision_keep") if isinstance(prof, dict) else None, 5
-            ),
-            failed_max_age_days=_coerce_int(
-                que.get("failed_max_age_days") if isinstance(que, dict) else None, 30
-            ),
-            failed_max_files=_coerce_int(
-                que.get("failed_max_files") if isinstance(que, dict) else None, 500
-            ),
-            failed_cleanup_interval=_coerce_int(
-                que.get("failed_cleanup_interval") if isinstance(que, dict) else None,
-                100,
-            ),
-            job_max_retries=_coerce_int(
-                que.get("job_max_retries") if isinstance(que, dict) else None, 3
-            ),
-        )
-
-    @staticmethod
-    def _parse_memes_config(data: dict[str, Any]) -> MemeConfig:
-        section_raw = data.get("memes", {})
-        section = section_raw if isinstance(section_raw, dict) else {}
-        return MemeConfig(
-            enabled=_coerce_bool(section.get("enabled"), True),
-            query_default_mode=_coerce_str(section.get("query_default_mode"), "hybrid"),
-            max_source_image_bytes=max(
-                1,
-                _coerce_int(section.get("max_source_image_bytes"), 500 * 1024),
-            ),
-            blob_dir=_coerce_str(section.get("blob_dir"), "data/memes/blobs"),
-            preview_dir=_coerce_str(section.get("preview_dir"), "data/memes/previews"),
-            db_path=_coerce_str(section.get("db_path"), "data/memes/memes.sqlite3"),
-            vector_store_path=_coerce_str(
-                section.get("vector_store_path"), "data/memes/chromadb"
-            ),
-            queue_path=_coerce_str(section.get("queue_path"), "data/memes/queues"),
-            max_items=max(1, _coerce_int(section.get("max_items"), 10000)),
-            max_total_bytes=max(
-                1,
-                _coerce_int(section.get("max_total_bytes"), 5 * 1024 * 1024 * 1024),
-            ),
-            allow_gif=_coerce_bool(section.get("allow_gif"), True),
-            auto_ingest_group=_coerce_bool(section.get("auto_ingest_group"), True),
-            auto_ingest_private=_coerce_bool(section.get("auto_ingest_private"), True),
-            keyword_top_k=max(1, _coerce_int(section.get("keyword_top_k"), 30)),
-            semantic_top_k=max(1, _coerce_int(section.get("semantic_top_k"), 30)),
-            rerank_top_k=max(1, _coerce_int(section.get("rerank_top_k"), 20)),
-            worker_max_concurrency=max(
-                1, _coerce_int(section.get("worker_max_concurrency"), 4)
-            ),
-        )
-
-    @staticmethod
-    def _parse_api_config(data: dict[str, Any]) -> APIConfig:
-        section_raw = data.get("api", {})
-        section = section_raw if isinstance(section_raw, dict) else {}
-
-        enabled = _coerce_bool(section.get("enabled"), True)
-        host = _coerce_str(section.get("host"), DEFAULT_API_HOST)
-        port = _coerce_int(section.get("port"), DEFAULT_API_PORT)
-        if port <= 0 or port > 65535:
-            port = DEFAULT_API_PORT
-
-        auth_key = _coerce_str(section.get("auth_key"), DEFAULT_API_AUTH_KEY)
-        if not auth_key:
-            auth_key = DEFAULT_API_AUTH_KEY
-
-        openapi_enabled = _coerce_bool(section.get("openapi_enabled"), True)
-
-        tool_invoke_enabled = _coerce_bool(section.get("tool_invoke_enabled"), False)
-        tool_invoke_expose = _coerce_str(
-            section.get("tool_invoke_expose"), "tools+toolsets"
-        )
-        valid_expose = {"tools", "toolsets", "tools+toolsets", "agents", "all"}
-        if tool_invoke_expose not in valid_expose:
-            tool_invoke_expose = "tools+toolsets"
-        tool_invoke_allowlist = _coerce_str_list(section.get("tool_invoke_allowlist"))
-        tool_invoke_denylist = _coerce_str_list(section.get("tool_invoke_denylist"))
-        tool_invoke_timeout = _coerce_int(section.get("tool_invoke_timeout"), 120)
-        if tool_invoke_timeout <= 0:
-            tool_invoke_timeout = 120
-        tool_invoke_callback_timeout = _coerce_int(
-            section.get("tool_invoke_callback_timeout"), 10
-        )
-        if tool_invoke_callback_timeout <= 0:
-            tool_invoke_callback_timeout = 10
-
-        return APIConfig(
-            enabled=enabled,
-            host=host,
-            port=port,
-            auth_key=auth_key,
-            openapi_enabled=openapi_enabled,
-            tool_invoke_enabled=tool_invoke_enabled,
-            tool_invoke_expose=tool_invoke_expose,
-            tool_invoke_allowlist=tool_invoke_allowlist,
-            tool_invoke_denylist=tool_invoke_denylist,
-            tool_invoke_timeout=tool_invoke_timeout,
-            tool_invoke_callback_timeout=tool_invoke_callback_timeout,
-        )
-
-    @staticmethod
-    def _parse_naga_config(data: dict[str, Any]) -> NagaConfig:
-        section_raw = data.get("naga", {})
-        section = section_raw if isinstance(section_raw, dict) else {}
-
-        enabled = _coerce_bool(section.get("enabled"), False)
-        api_url = _coerce_str(section.get("api_url"), "")
-        api_key = _coerce_str(section.get("api_key"), "")
-        moderation_enabled = _coerce_bool(section.get("moderation_enabled"), True)
-        allowed_groups = frozenset(_coerce_int_list(section.get("allowed_groups")))
-
-        return NagaConfig(
-            enabled=enabled,
-            api_url=api_url,
-            api_key=api_key,
-            moderation_enabled=moderation_enabled,
-            allowed_groups=allowed_groups,
-        )
-
-    @staticmethod
-    def _parse_easter_egg_call_mode(value: Any) -> str:
-        """解析彩蛋提示模式。
-
-        兼容旧版布尔值：
-        - True  => agent
-        - False => none
-        """
-        if isinstance(value, bool):
-            return "agent" if value else "none"
-        if isinstance(value, (int, float)):
-            return "agent" if bool(value) else "none"
-        if value is None:
-            return "none"
-
-        text = str(value).strip().lower()
-        if text in {"true", "1", "yes", "on"}:
-            return "agent"
-        if text in {"false", "0", "no", "off"}:
-            return "none"
-        if text in {"none", "agent", "tools", "all", "clean"}:
-            return text
-        return "none"
 
     def reload(self, strict: bool = False) -> dict[str, tuple[Any, Any]]:
         new_config = Config.load(strict=strict)
@@ -3346,20 +1594,3 @@ class Config:
 
     def is_admin(self, qq: int) -> bool:
         return qq in self.admin_qqs
-
-
-def _update_dataclass(
-    old_value: Any, new_value: Any, prefix: str
-) -> dict[str, tuple[Any, Any]]:
-    changes: dict[str, tuple[Any, Any]] = {}
-    if not isinstance(old_value, type(new_value)):
-        changes[prefix] = (old_value, new_value)
-        return changes
-    for field in fields(old_value):
-        name = field.name
-        old_attr = getattr(old_value, name)
-        new_attr = getattr(new_value, name)
-        if old_attr != new_attr:
-            setattr(old_value, name, new_attr)
-            changes[f"{prefix}.{name}"] = (old_attr, new_attr)
-    return changes
